+++ Engine/JocysCom/Controls/Chat/ChatSettings.cs
@@ -5,6 +5,6 @@ namespace JocysCom.ClassLibrary.Controls.Chat
 	public class ChatSettings
 	{
 		[DefaultValue(0)]
-		public int ScrollPosition { get;set; }
+		public int ScrollPosition { get; set; }
 	}
 }